# backend/static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Categorys": {
				"properties": {
					"categoryName": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"categoryName",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"keycloak": {
				"type": "apiKey",
				"name": "Authorization",
				"description": "Keycloak Bearer Token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/Category": {
			"get": {
				"operationId": "GetCategoryAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"CategoryName": {
												"type": "string"
											},
											"Description": {
												"type": "string"
											},
											"Id": {
												"type": "string"
											}
										},
										"required": [
											"CategoryName",
											"Description",
											"Id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Categorys"
							}
						}
					}
				}
			}
		},
		"/Category/{id}": {
			"patch": {
				"operationId": "UpdataCategory",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Categorys"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Delete",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}